+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/package.json
@@ -7,6 +7,7 @@
         "build": "vue-cli-service build"
     },
     "dependencies": {
+        "axios": "^0.27.2",
         "bootstrap": "^5.1.3",
         "core-js": "^3.8.3",
         "popper.js": "^1.16.1",
