+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/src/api/core/prepareRequest.js
@@ -1,0 +1,30 @@
+export default pl => {
+    const  { method= 'GET' , url , parametres = {} } = pl;
+    const base = { method, url };
+
+    switch (method) {
+        case 'POST': {
+            const { data = {}, params = {}} = parametres;
+            base.data = data;
+            base.params = params;
+            break;
+        }
+        case 'PUT': {
+            const { data = {}, params = {}} = parametres;
+            base.data = data;
+            base.params = params;
+            break;
+        }
+        case 'DELETE': {
+            const { data = {}, params = {}} = parametres;
+            base.data = data;
+            base.params = params;
+            break;
+        }
+        default: {
+            //GET
+            base.params = parametres;
+        }
+    }
+    return base;
+};
