+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/src/api/core/prepareResponse.js
@@ -1,0 +1,8 @@
+export default d => {
+    const { data } = d;
+    if (!data) {
+        throw new Error();
+    } else {
+        return data;
+    }
+};
