+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/src/api/index.js
@@ -1,0 +1,1 @@
+export * from './modules';
