+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/src/api/modules/catalog.js
@@ -1,0 +1,15 @@
+import instance  from '@api/core';
+
+const url = '/catalog';
+
+export const catalog  = {
+    async getCatalog() {
+        try{
+            const data = await instance({url});
+            return data;
+        }
+        catch(err) {
+            throw err;
+        }
+    },
+};
