+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/src/api/modules/index.js
@@ -1,0 +1,7 @@
+import { catalog } from "./catalog";
+import { cart } from './cart';
+
+export {
+    catalog,
+    cart,
+};
