+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/src/components/Header.vue
@@ -112,11 +112,6 @@ export default {
    components: {
     Cart
   },
-  methods: {
-    hi(){
-      console.log(this.item)
-    }
-  },
 }
 </script>
