+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/src/components/items/cartItem.vue
