+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/src/componentsPage/homeComponents/iconBoxes/iconIcon/IconItems.vue
@@ -17,10 +17,6 @@
 export default {
   props: ['item' , 'img'],
   name: 'IconItems',
-
-created() {
-  console.log(this.img + this.item.imgUrl)
-}
 };
 </script>
