+++ STUDENTS/Trofimov/trofimov_hw_VUE_6/project_shop/src/store/modules/catalog.js
@@ -1,3 +1,5 @@
+import  { catalog }  from "@api";
+
 export default {
     namespaced: true,
     state: () => ({
@@ -5,9 +7,9 @@ export default {
     }),
 
     actions: {
-        async getCatalog({ commit }, url) {
+        async getCatalog({ commit }) {
             try {
-                const data = await $api.send(url, 'GET');
+                const data = await catalog.getCatalog();
                 commit('setCatalogItems', data);
             } catch(err) {
                 throw err;
